WIP: Move dash_info.dart and reddit_video_metadata.dart to values/reddit/
- Classes are value objects rather than models.
- Update references to reflect movement.
